Update code in `nl2sql.ipynb` notebook
Changed the approach to explain the SQL tables to the model, using a domain specific language already understood by the model.
This results in a more compact system prompt that still allows for a more nuanced description of the data model.
This should result in less tokens consumed by the system prompts and better maintainability allowing for a richer data definition.