Update Prompt Engineering OpenAI notebook
The main change in the update was adjusting and correcting a few typos, fixing some grammar errors, translating to English a text that was in Spanish, and fixing a driver's last name, which had a spelling error.
  - Introducction -> Introducction
  - a API -> an API
  - Formating -> Formatting
  - Functio -> Function
  - "A un modelo tan grande y bueno como GPT 3.5 le basta un solo shot para aprende el formato de salida que esperamos." -> "For a model as large and good as GPT 3.5, a single shot is enough to learn the output format we expect."
  - Bettel -> Vettel
  - answew -> answer
  - Setiment -> Sentiment

This improves the readability and comprehension of the document.
The update does not alter the main code or results.